09/12/23

+++ backend/src/main/java/br/com/docvet/docvet/service/VeterinarioService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import br.com.docvet.docvet.domain.Veterinario;
+import br.com.docvet.docvet.domain.error.NotFoundException;
 import br.com.docvet.docvet.repository.VeterinarioRepository;
 
 @Service
@@ -15,7 +16,7 @@ public class VeterinarioService {
     private VeterinarioRepository repository;
 
     public Veterinario getOne(Integer id){
-        return repository.findById(id).get();
+        return repository.findById(id).orElseThrow(()-> new NotFoundException("Veterinário não encontrado: id="+id));
     }
 
     public List<Veterinario> getAll(){
@@ -27,18 +28,25 @@ public class VeterinarioService {
     }
 
     public void update(Veterinario veterinario, Integer id){
-        Veterinario existente = repository.findById(id).get();
+        if (!repository.existsById(veterinario.getId())) {
+            throw new NotFoundException("Veterinário não encontrado: id="+veterinario.getId());
 
-        existente.setNome(veterinario.getNome());
-        existente.setEmail(veterinario.getEmail());
-        existente.setFoto(veterinario.getFoto());
-        existente.setCrmv(veterinario.getCrmv());
+        }
+        // Veterinario existente = repository.findById(id).get();
 
-        repository.saveAndFlush(existente);
+        // existente.setNome(veterinario.getNome());
+        // existente.setEmail(veterinario.getEmail());
+        // existente.setFoto(veterinario.getFoto());
+        // existente.setCrmv(veterinario.getCrmv());
+
+        repository.saveAndFlush(veterinario);
 
     }
 
     public void delete(Integer id){
+        if (!repository.existsById(id)) {
+            throw new NotFoundException("Veterinário não encontrado: id="+id);
+        }
         repository.deleteById(id);
     }
 }
